olaru-bot/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
agno[postgres]
groq==0.13.0
redis==5.2.1
psycopg2-binary==2.9.10
requests==2.32.3
structlog==24.4.0
python-dotenv==1.0.1
bcrypt==4.2.1
google-generativeai==0.8.3
